services/auth-service/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
cryptography==42.0.2
Pillow==10.2.0
celery==5.3.6
drf-spectacular==0.27.1
django-cors-headers==4.3.1
phonenumbers==8.13.29
gunicorn==21.2.0
